ML_Pipeline/tests/test_backward_compatibility.py
"""
BACKWARD COMPATIBILITY TEST
Verifies that enterprise upgrade doesn't break anything
"""
# ML-Pipeline/test_backward_compatibility.py

#!/usr/bin/env python3
"""
Backward Compatibility Test
Runs comprehensive tests to ensure zero breaking changes
"""

import sys
import os
from pathlib import Path
import subprocess
import json
import yaml
import logging

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

def test_original_pipeline():
    """Test that original pipeline still works"""
    print("\n🧪 TEST 1: Original Pipeline Compatibility")
    print("-" * 40)
    
    # Check if original main.py exists
    main_path = Path("src/main.py")
    if not main_path.exists():
        print("❌ src/main.py not found")
        return False
    
    print("✓ Original src/main.py exists")
    
    # Check if original config exists
    config_path = Path("config/ml_config.yaml")
    if not config_path.exists():
        print("❌ config/ml_config.yaml not found")
        return False
    
    print("✓ Original config exists")
    
    # Check if we can import original modules
    try:
        sys.path.insert(0, str(Path(__file__).parent))
        from src.data_collector import DataCollector
        from src.feature_engineer import FeatureEngineer
        from src.ensemble_model import TradingEnsembleModel
        print("✓ Original modules can be imported")
    except ImportError as e:
        print(f"❌ Could not import original modules: {e}")
        return False
    
    return True

def test_ai_service_compatibility():
    """Test that AI-Service can still load models"""
    print("\n🧪 TEST 2: AI-Service Compatibility")
    print("-" * 40)
    
    # Check models directory exists
    models_dir = Path("models")
    if not models_dir.exists():
        print("⚠️  models/ directory doesn't exist (will be created on first training)")
        models_dir.mkdir(exist_ok=True)
    
    # Check enterprise config points to correct location
    enterprise_config = Path("config/enterprise_config.yaml")
    if enterprise_config.exists():
        with open(enterprise_config, 'r') as f:
            config = yaml.safe_load(f)
        
        model_path = config.get('backward_compatibility', {}).get('model_path', 'models/')
        if model_path == 'models/':
            print("✓ Enterprise config points to correct model path")
        else:
            print(f"⚠️  Enterprise config model path: {model_path}")
    else:
        print("⚠️  Enterprise config not found")
    
    # Check AI-Service config (simulated)
    ai_service_config_content = """
ml:
  model_path: "../ML-Pipeline/models/production/"
  models:
    - name: "ensemble"
      file: "ensemble_model_SOL_USDT.joblib"
"""
    
    print("✓ AI-Service config would load from: ../ML-Pipeline/models/production/")
    print("  Note: This is a symlink to registry/production/sol_usdt/current/")
    
    return True

def test_model_registry():
    """Test that model registry works with backward compatibility"""
    print("\n🧪 TEST 3: Model Registry & Backward Compatibility")
    print("-" * 40)
    
    try:
        sys.path.insert(0, str(Path(__file__).parent))
        from enterprise.model_registry import ModelRegistry, ModelStatus
        
        # Initialize registry
        registry = ModelRegistry()
        print("✓ ModelRegistry initialized")
        
        # Check directories created
        if Path("registry").exists():
            print("✓ registry/ directory exists")
        else:
            print("❌ registry/ directory not created")
            return False
        
        # Check production directory
        if Path("registry/production").exists():
            print("✓ registry/production/ directory exists")
        else:
            print("❌ registry/production/ directory not created")
            return False
        
        # Check legacy models directory
        if Path("models").exists():
            print("✓ models/ directory exists (for backward compatibility)")
        else:
            print("⚠️  models/ directory doesn't exist")
        
        return True
        
    except Exception as e:
        print(f"❌ ModelRegistry test failed: {e}")
        return False

def test_enterprise_orchestrator():
    """Test that enterprise orchestrator can be initialized"""
    print("\n🧪 TEST 4: Enterprise Orchestrator")
    print("-" * 40)
    
    try:
        sys.path.insert(0, str(Path(__file__).parent))
        from enterprise.orchestrator import EnterpriseOrchestrator
        
        # Try to initialize
        orchestrator = EnterpriseOrchestrator()
        print("✓ EnterpriseOrchestrator initialized")
        
        # Check it has original pipeline
        if hasattr(orchestrator, 'original_pipeline'):
            print("✓ Has reference to original MLPipeline")
        else:
            print("⚠️  No reference to original MLPipeline")
        
        # Check registry
        if hasattr(orchestrator, 'registry'):
            print("✓ Has ModelRegistry instance")
        else:
            print("⚠️  No ModelRegistry instance")
        
        return True
        
    except Exception as e:
        print(f"❌ EnterpriseOrchestrator test failed: {e}")
        return False

def test_symlink_creation():
    """Test that symlinks are created for backward compatibility"""
    print("\n🧪 TEST 5: Symlink Backward Compatibility")
    print("-" * 40)
    
    # Create a test model in registry
    test_registry_dir = Path("registry/production/test_symbol/v1.0.0")
    test_registry_dir.mkdir(parents=True, exist_ok=True)
    
    # Create dummy model file
    dummy_model = test_registry_dir / "model.joblib"
    import joblib
    joblib.dump({"test": "model"}, dummy_model)
    
    # Create current symlink
    current_link = Path("registry/production/test_symbol/current")
    if current_link.exists():
        current_link.unlink()
    current_link.symlink_to(test_registry_dir, target_is_directory=True)
    
    # Create backward compatibility symlink
    legacy_path = Path("models/ensemble_model_test_symbol.joblib")
    if legacy_path.exists():
        legacy_path.unlink()
    
    try:
        legacy_path.symlink_to(dummy_model)
        print("✓ Can create backward compatibility symlinks")
        
        # Clean up
        legacy_path.unlink(missing_ok=True)
        import shutil
        shutil.rmtree("registry/production/test_symbol", ignore_errors=True)
        
        return True
    except Exception as e:
        print(f"❌ Symlink creation failed: {e}")
        return False

def run_all_tests():
    """Run all backward compatibility tests"""
    print("=" * 60)
    print("BACKWARD COMPATIBILITY TEST SUITE")
    print("=" * 60)
    
    tests = [
        ("Original Pipeline", test_original_pipeline),
        ("AI-Service Compatibility", test_ai_service_compatibility),
        ("Model Registry", test_model_registry),
        ("Enterprise Orchestrator", test_enterprise_orchestrator),
        ("Symlink Creation", test_symlink_creation)
    ]
    
    results = []
    
    for test_name, test_func in tests:
        try:
            success = test_func()
            results.append((test_name, success))
        except Exception as e:
            print(f"❌ {test_name} test crashed: {e}")
            results.append((test_name, False))
    
    # Summary
    print("\n" + "=" * 60)
    print("TEST SUMMARY")
    print("=" * 60)
    
    passed = sum(1 for _, success in results if success)
    total = len(results)
    
    for test_name, success in results:
        status = "✅ PASS" if success else "❌ FAIL"
        print(f"{status} - {test_name}")
    
    print(f"\nTotal: {passed}/{total} tests passed")
    
    if passed == total:
        print("\n🎉 ALL TESTS PASSED!")
        print("Enterprise upgrade maintains 100% backward compatibility!")
        return True
    else:
        print(f"\n⚠️  {total - passed} tests failed")
        print("Some backward compatibility issues detected")
        return False

if __name__ == "__main__":
    success = run_all_tests()
    sys.exit(0 if success else 1)
