ML_Pipeline/requirements.txt
# Core ML
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
catboost==1.2.4
tensorflow==2.15.0  # For neural networks

# Data Processing
pandas==2.1.4
numpy==1.24.4
pandas-ta==0.3.14  # Technical indicators
ta==0.10.2        # Alternative TA library
yfinance==0.2.33  # Market data
ccxt==4.1.52      # Crypto exchange data

# Feature Engineering
tsfresh==0.20.2   # Time series features
feature-engine==1.6.2

# Reinforcement Learning
gym==0.26.2
stable-baselines3==2.1.0
torch==2.1.0

# Visualization
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.0

# Utilities
joblib==1.3.2
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
